requirements.txt
flask==2.2.3
flask_restx==1.0.6
celery==5.2.7
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-Celery-Helper==1.1.0
gunicorn==20.1.0
gevent==22.10.2
paramiko==3.0.0
pywinrm==0.4.3
requests==2.27.1
retrying==1.3.3
sentry-sdk[flask]==1.24.0
pyodbc==4.0.35
cx-Oracle==7.3.0
psycopg2-binary==2.9.5
PyMySQL==0.9.3
py2neo==2021.2.3
Cython==0.29.32
happybase==1.2.0
thrift==0.16.0
confluent-kafka==2.1.1
jmxquery==0.6.0
kazoo==2.8.0
dmPython==2.3
pymongo==4.3.3
xmltodict==0.12.0
cryptography==3.3.1
telnetlib3==1.0.4
pyvmomi==7.0.3
grequests==0.6.0
ujson==1.34
chardet==3.0.4
info2soft==0.1.1
packaging==21.3
pymssql==2.2.7
websocket-client==1.3.1
pyhdb==0.3.4
kubernetes==24.2.0
rpyc==5.0.1
dictdiffer==0.9.0